Make Scripts module structure
