+++ app/helpers/scripts/wiki_igrapher/schema/input.yml
